### 2022-04-27 #4351 @editorialbot
[PRE REVIEW]:  elwaspatid: A Python package to compute space-time diagrams for the propagation of elastic waves in 1D rods
**Submitting author:** @dbrizard (<a href="http://orcid.org/0000-0003-2264-7131">Denis Brizard</a>)
**Repository:** https://github.com/dbrizard/elwaspatid
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.1
