### 2022-04-27 #4351 @dbrizard
@danielskatz I think this would be useful to researchers in mechanical engineering:
- to simulate mechanisms in which we have contact/impact between translational parts;
- mechanisms which involve hammers (eg pile driving).

This would also be useful to researchers working, as I do, on Split Hopkinson Pressure Bars:
- to understand the wave propagation in the bars;
- to design a striker geometry in order to get a specific pulse shape.

I also believe this can help students understand the wave propagation phenomena (reflection and transmission) across section changes in bars.

---

I know 1D models are not the trend, but I think they allow a better understanding of the physics, and can be used in a fast pre-design stage.

I hope this answers your question.
