+++ setup.cfg
@@ -17,7 +17,7 @@ classifiers =
     Intended Audience :: Science/Research
     Topic :: Scientific/Engineering
     Topic :: Scientific/Engineering :: Physics
-keywords = wave propagation, elastic wave, SHPB, rod, 1D, traction, compression
+keywords = wave propagation, elastic wave, SHPB, rod, 1D, traction, compression, Lagrangian x-t diagram
 
 [options]
 package_dir =
