### 2022-05-17 097e842
energy computation
