+++ src/elwaspatid/elwaspatid.py
@@ -125,6 +125,7 @@ class WP2:
         for ss in bar.seg:
             ss.setTime(time) #set :attr:`time` for each :class:`Segment`
             ss.computeStressStrain()
+            ss.computeEnergy()
         
         self.time = time
         self.bar = bar
@@ -470,7 +471,31 @@ class WP2:
         plt.xlabel('t [s]')
         plt.ylabel('x [m]')
         plt.box(on=False)
+    
+    def plotEnergy(self, figname=None, ls='.-'):
+        """
 
+        :param str figname: name for the figure
+        :param str ls: linestyle
+        """
+        plt.figure(figname)
+        for ii, ss in enumerate(self.bar.seg):
+            if ii==0:
+                ax = plt.subplot(self.bar.nseg, 1, ii+1)
+            else:
+                plt.subplot(self.bar.nseg, 1, ii+1, sharex=ax)
+                
+            for kk in ss.energy:
+                plt.plot(ss.time, ss.energy[kk], ls, label=kk)
+            # plt.plot(ss.time, ss.energy['kinetic'], '.--', label=ii)
+            # plt.plot(ss.time, ss.energy['potential'], '.:', label=ii)
+            # plt.plot(ss.time, ss.energy['total'], '.-', label=ii)
+            if ii==0:
+                plt.legend()
+        plt.xlabel('time [s]')
+        plt.ylabel('energy [J]')
+        
+        
 class Waveprop:
     '''One-dimensional wave propagation problem for a rod with a piecewise
     constant impedance.
@@ -606,6 +631,13 @@ class Waveprop:
         self._Stress['right'] = self.Force[:,1:]/bar.A  # right stress, @elements
         # This should rather be the way
         self.Stress = self.Strain*bar.E
+        
+        # Kinetic and potential energy
+        # Ec = 0.5*np.sum(self.Veloc[:,:-1]**2*bar.A*bar.rho*np.diff(bar.x) , axis=1) # wrong!
+        # Ep = 0.5*np.sum(self.Force[:,:-1]**2/bar.E/bar.A*np.diff(bar.x), axis=1) # wring!
+        Ecm = 0.5*np.sum(self.Veloc[:,:-1]*self.Veloc[:,1:]*bar.A*bar.rho*np.diff(bar.x) , axis=1)
+        Epm = 0.5*np.sum(self.Force[:,:-1]*self.Force[:,1:]/bar.E/bar.A*np.diff(bar.x), axis=1)
+        self.energy = {'kinetic':Ecm, 'potential':Epm, 'total':Ecm+Epm}
         
         # Traction-Compression state
         LR = Force*Veloc
@@ -826,6 +858,23 @@ class Waveprop:
         plt.plot(ab, displ, 'g.-', drawstyle='steps-post')
         plt.ylabel('Displacement')
         plt.xlabel(xlab)
+    
+    
+    def plotEnergy(self, figname=None, ls='.-'):
+        """
+        
+        :param str figname: name for the figure
+        :param str ls: linestyle
+        """
+        plt.figure(figname)
+        # plt.plot(self.time, self.energy['kin'], ls, label='kinetic')
+        # plt.plot(self.time, self.energy['pot'], ls, label='potential')
+        # plt.plot(self.time, self.energy['kin']+self.energy['pot'], ls, label='total')
+        for kk in self.energy:
+            plt.plot(self.time, self.energy[kk], ls, label=kk)
+        plt.legend()
+        plt.xlabel('time [s]')
+        plt.ylabel('energy [J]')
         
     
     def plotFV(self, x, figname=None):
@@ -1105,7 +1154,7 @@ class BarSet(BarSingle):
[…]
