+++ docs/auto_examples/plot_1_WP2.rst
@@ -115,12 +115,13 @@ Contact interface between the two bars:
 
 And so on. Which means the pulse is trapped in the second bar.
 
-.. GENERATED FROM PYTHON SOURCE LINES 49-52
+.. GENERATED FROM PYTHON SOURCE LINES 49-53
 
 .. code-block:: default
 
     test2 = WP2(bc, comp, nstep=100, left='free', right='free')
     test2.plot('2b_free')
+    test2.plotEnergy()
 
 
 
@@ -149,21 +150,36 @@ And so on. Which means the pulse is trapped in the second bar.
          :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_003.png
          :class: sphx-glr-multi-img
 
+    *
+
+      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_004.png
+         :alt: energy [J]
+         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_004.png
+         :class: sphx-glr-multi-img
+
+    *
+
+      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_005.png
+         :alt: plot 1 WP2
+         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_005.png
+         :class: sphx-glr-multi-img
 
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 53-55
+
+.. GENERATED FROM PYTHON SOURCE LINES 54-56
 
 Two identical bars, free and fixed ends
 ---------------------------------------
 
-.. GENERATED FROM PYTHON SOURCE LINES 55-58
+.. GENERATED FROM PYTHON SOURCE LINES 56-60
 
 .. code-block:: default
 
     test2 = WP2(bc, comp, nstep=100, left='free', right='fixed')
     test2.plot('2b_freefixed')
+    test2.plotEnergy()
 
 
 
@@ -173,30 +189,44 @@ Two identical bars, free and fixed ends
 
     *
 
-      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_004.png
-         :alt: Force [N]
-         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_004.png
-         :class: sphx-glr-multi-img
-
-    *
-
-      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_005.png
-         :alt: Velocity [m/s]
-         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_005.png
-         :class: sphx-glr-multi-img
-
-    *
-
       .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_006.png
-         :alt: Displacement [m]
+         :alt: Force [N]
          :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_006.png
          :class: sphx-glr-multi-img
 
+    *
+
+      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_007.png
+         :alt: Velocity [m/s]
+         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_007.png
+         :class: sphx-glr-multi-img
+
+    *
+
+      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_008.png
+         :alt: Displacement [m]
+         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_008.png
+         :class: sphx-glr-multi-img
+
+    *
+
+      .. image-sg:: /auto_examples/images/sphx_glr_plot_1_WP2_009.png
+         :alt: energy [J]
+         :srcset: /auto_examples/images/sphx_glr_plot_1_WP2_009.png
[…]
