### 2022-05-20 #4351 @erasmuss
Hi,

Sorry for the late reply, works been hectic and I’m traveling, I can review
it next week, already started reading the page.

Please confirm this.

Thanks,
Elizabeth ***@***.***)

On Fri, May 20, 2022 at 8:37 AM Patrick Diehl ***@***.***>
wrote:

> Hi @MargeritPierre <https://github.com/MargeritPierre> or @ntlva
> <https://github.com/ntlva> would you be interested in reviewing this
> paper?
>
