+++ docs/auto_examples/plot_1_WP2.rst
@@ -604,7 +604,7 @@ Nothing happens, the left bar travels to the left.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  13.723 seconds)
+   **Total running time of the script:** ( 0 minutes  14.302 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_1_WP2.py:
