### 2022-05-26 #4351 @diehlpk
> Hi, Sorry for the late reply, works been hectic and I’m traveling, I can review it next week, already started reading the page. Please confirm this. Thanks, Elizabeth ***@***.***)
> On Fri, May 20, 2022 at 8:37 AM Patrick Diehl ***@***.***> wrote: Hi @MargeritPierre <https://github.com/MargeritPierre> or @ntlva <https://github.com/ntlva> would you be interested in reviewing this paper? — Reply to this email directly, view it on GitHub <[#4351 (comment)](https://github.com/openjournals/joss-reviews/issues/4351#issuecomment-1132975543)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AJVSCDFT4J26GDWDA4HPU6TVK6PTXANCNFSM5UOTW53A> . You are receiving this because you were mentioned.Message ID: ***@***.***>
> -- Elizabeth Rasmussen

@erasmuss sounds good. Sorry for the late respond, I was on vacation.
