+++ paper/paper.md
@@ -22,8 +22,9 @@ bibliography: paper.bib
 the computation of space-time diagrams. It is meant to model the propagation of
 elastic stress waves in slender bars -or rods- generated by the impact of another bar. 
 Application can be: mechanical testing (Split Hopkinson Pressure Bars), 
- mechanical systems which involve impact with a hammer, understanding of wave 
-reflection/transmission on academical cases.
+ mechanical systems which involve impact with a hammer, design of striker 
+geometry to get a specific pulse shape, understanding of wave 
+reflection/transmission in bars on academical cases.
 
 
 # Statement of Need
