+++ docs/conf.py
@@ -34,6 +34,7 @@ extensions = [
     'sphinx.ext.autodoc', # to get automodule. https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html
     'sphinx_gallery.gen_gallery',
     'sphinx.ext.intersphinx',
+    'sphinx_rtd_theme',
 #    "sphinx_codeautolink", # not working either
 ]
 
@@ -70,7 +71,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'  # 'alabaster'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
