### 2022-05-30 #4418 @obsmax
The LICENCE file contains the text of the GNU-GPL 3.0 licence but the setup.cfg file indicates that the licence is a "MIT License".
