### 2022-05-30 #4418 @obsmax
- Installation
The [installation instructions](https://elwaspatid.readthedocs.io/en/latest/getting-started.html) recommend using `pip install elwaspatid` but this command do not bring the `examples` folder that is required to test the installation. To run the examples, the user is invited to get it from [github](https://github.com/dbrizard/elwaspatid) or from the [doc](https://elwaspatid.readthedocs.io/en/latest/getting-started.html).
The issue is that, the user might be trying to run examples corresponding to a version of the program that do not match with the version gathered and installed from pypi.
A simple fix would be to add a `setup.py` file to the package so that users could install the package locally directly from the source directory using `pip install -e .` after `git clone`.
Users might then be guided to PyPi for applications only or to the Github sources if they want to run the examples and learn how to use it.
