### 2022-05-30 #4418 @obsmax
- Version tag : I'm not sure that the version tag `v2.0.1` complies with the PEP440 rule,  shouldn't it be `2.0.1` ?
