### 2022-05-30 #4418 @obsmax
- Ideally, the dependencies (`numpy`, `matplotlib`[, `jupyter notebook`]) might be listed in a requirement file or being installed with the package, through the pip installation command.
