### 2022-05-30 #4418 @obsmax
- Test examples
  - `plot_2_ElasticImpact.py`: a warning message suggests that things are going wrong : `ça va pas !!`, is it the case?
  - The file description of `examples/plot_6_underHood.py` differs slightly from `docs/auto_examples/plot_6_underHood.py`.    Symlinks might avoid these files to diverge in future versions.
  - The examples programs generate a lot of figures but do not display them (the `plt.show()` instruction is missing).
    If the main purpose of these examples is to test the installation,  then unit-tests called with pytest might be more adapted.
