### 2022-05-30 #4418 @obsmax
- The sources still include old programs (e.g. `prop1D_old.py`). The git history can be used to easily recover old programs, while keeping the sources clean.
