### 2022-05-31 #4418 @dbrizard
Thank you @obsmax for pointing out these issues.
As you might guess, this is my first serious git repository and Python package, and I am still learning !
I will try to correct them all.

- [ ] licence should be the same everywhere
- [ ] installation with `setup.py` file
- [ ] check version tag with `v` is ok or not
- [ ] handle dependencies automatically
- [ ] check examples
- [ ] remove old programs (Indeed, this was before I used git...)
