### 2022-05-31 #4418 @dbrizard
@obsmax #https://github.com/openjournals/joss-reviews/issues/4418#issuecomment-1141469926

# test examples

- warning message in example 2 was obsolete (analytical and numerical solutions overlay, see [last figure](https://elwaspatid.readthedocs.io/en/latest/auto_examples/plot_2_ElasticImpact.html#compare-analytical-and-numerical-solutions)) . Removed.
- files in `docs/auto_examples/` are copies from `examples/`, this is automatically done by *sphinx_gallery* extension. I will take care to `make clean` before `make html` when building *sphinx* documentation.
- `plt.show()` added in the example files.

>  If the main purpose of these examples is to test the installation, then unit-tests called with pytest might be more adapted.

The purpose of the examples is threefold:

1. show how to use the module;
2. display several academic cases of bar configurations (educational purpose);
3. visually check that the space-time diagrams on these academic cases are correct.

I considered writing unit-tests, but struggled on the  `assert` statement to use since the main outputs of the module are large arrays (force or velocity wrt time and space). That is why I went for *sphinx_gallery*.
