+++ examples/plot_4_SHPB.py
@@ -39,3 +39,4 @@ plt.plot(testk.time*1e6, -f2/1000, '-', label='right')
 plt.xlabel('time [µs]')
 plt.ylabel('force [kN]')
 plt.legend()
+plt.show()
