+++ src/elwaspatid/prop1D_old.py
@@ -1,1422 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-This module is based on the following paper:
-
-Bacon, C. (1993). Numerical prediction of the propagation of elastic waves in 
-longitudinally impacted rods : Applications to Hopkinson testing. 
-*International Journal of Impact Engineering*, 13(4), 527‑539. 
-https://doi.org/10.1016/0734-743X(93)90084-K
-
-Summary-- Simple expressions, based on one-dimensional elastic wave theory, are
-established which permit prediction of normal force and particle velocity at 
-cross-sections of a non-uniform linearly-elastic rod. The initial normal force
-and particle velocity at each cross-section of that rod must be known.
-
-
-First define a bar, then give it to :class:`WP2` or :class:`Waveprop` along 
-with an incident wave for propagation computation::
-
-    bar = Barhete(E=[210, 78], rho=[7800, 2800], L=[1, 1.1], d=[0.030, 0.028])
-    incw = np.ones(100)
-    prop = WP2(bar, incw)
-
-Be careful:
-
-* :class:`WP2` works only with :class:`Barhete` bars; traction is not transmitted throught interfaces
-* :class:`Waveprop` works with :class:`Barhete` and :class:`Barhomo` bars, but does not take interfaces between bars/segments into account (the bars are stuck, traction can cross interfaces)
-
-Created on Fri Aug 22 11:13:37 2014
-
-@author: dbrizard
-"""
-
-import numpy as np
-import matplotlib.pyplot as plt
-import warnings
-
-# import figutils as fu
-
-
-class WP2:
-    """Second version of wave propagation, using :class:`Segment` for each bar
-    of constant section.
-    
-    No traction crossing interfaces (=section changes).
-
-    
-    """
-    def __init__(self, bar, incw=None, nstep=0, left='free', right='free', Vinit=0):
-        """Computte wave propagation
-        
-        /!\ Anechoic condition at impact end (left) until the end of the 
-        prescribed incident wave *incw*
-        
-        :param obj bar: bar setup (:class:`Barhete` object)
-        :param array incw: incident force wave (input left impact)
-        :param int nstep: optional number of time step
-        :param str left: left boundary condition, once incident wave is finished
-        :param str right: right boundary condition ('free' or 'plain')
-        :param float Vinit: initial velocity of left bar
-        """
-        if nstep==0:
-            n_trav = 2 #number of wave travels through the entire bar
-            nstep = n_trav*np.sum(bar.nelt)
-            print("%i traversées de barres"%n_trav)
-        
-        nT = nstep #len(incw)
-
-        # Initial conditions: at rest (first line) + initialization of matrices
-        for ii, ss in enumerate(bar.seg):
-            if ii==0 and not Vinit==0:
-                print("Setting initial velocity of first segment (Vo=%g)"%Vinit)
-                ss.initCalc(nT, Vo=Vinit)
-                incw = np.zeros(0)
-            else:
-                ss.initCalc(nT)
-            
-        for it in range(nT)[1:]:
-            for ii, ss in enumerate(bar.seg):
-                ss.compMiddle(it) #middle state of each segment
-                if ii==0: 
-                    #first segment LEFT:
-                    if it<=len(incw):
-                        ss.compLeft(it, incw=incw[it-1]) # excited
-                    else:
-                        ss.compLeft(it, left=left) # not excited any more
-                    #first segment RIGHT:
[…]

[…]
