+++ src/elwaspatid.egg-info/SOURCES.txt
@@ -1,10 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-setup.cfg
-src/elwaspatid/__init__.py
-src/elwaspatid/elwaspatid.py
-src/elwaspatid.egg-info/PKG-INFO
-src/elwaspatid.egg-info/SOURCES.txt
-src/elwaspatid.egg-info/dependency_links.txt
-src/elwaspatid.egg-info/top_level.txt

[…]
