### 2022-06-01 b7c2864
docs: installation instructions
