### 2022-06-01 c95c35b
docs: plot_6 example
