### 2022-06-13 #4418 @dbrizard
@MargeritPierre thank you for the comments and the review.
Yes, there are no novelty in this program, apart from additional boundary conditions compared to the equations given by Bacon in his 1993 paper. Again, nothing extraordinary nor new. The purpose is to automatically draw space-time diagram, and be able to understand complex signals when multiple waves from various reflections overlay.

Indeed, I use it to understand what happens in SHPB devices. But with the limitation that the dispersion is not taken into account.
As you point it out, adding dispersion would be interesting. However, I believe this would require to heavily redesign the software because the underlying equations and numerical methods would be completely different. _I guess I would benefit from having a look at what is done in the seismic field..._

I tried to find similar software which plot space-time diagram, but was not able to find any.
I will add references to SHPB processing software, even if this kind of tool is rather dedicated to post-treat strain signal (dispersion correction, computation of stress and strain in the sample from the measured strain in the bars) and are generally not design tools.
