### 2022-06-14 85f64ed
docs: paper updated (state of field section)
