+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 `elwaspatid` is a Python package for 1D elastic wave propagation dedicated to 
 the computation of space-time diagrams. It is meant to model the propagation of
 elastic stress waves in slender bars -or rods- generated by the impact of another bar. 
-Application can be: mechanical testing (Split Hopkinson Pressure Bars), 
+Application can be: mechanical testing (Split Hopkinson Pressure Bars, SHPB), 
  mechanical systems which involve impact with a hammer, design of striker 
 geometry to get a specific pulse shape, understanding of wave 
 reflection/transmission in bars on academical cases.
@@ -44,6 +44,27 @@ shape is not modified by the propagation, but this approximation
 is rather acceptable in numerous cases, when the length of the wave is several
 times greater than the bar radius [@johnson_impact_1972]. The material is also 
 supposed to stay in the elastic range of deformation. 
+
+# State of the field
+
+Wave propagation softwares can be found in many scientific fields. 
+Williams et al [@williams_one-dimensional_1996] proposed a FORTRAN program for 
+the propagation of wavefronts in layered media, this instructional tool is based 
+on finite-difference and outputs the displacement with respect to the spatial variable. 
+Computer Programs in Seismology [@herrmann_computer_2013] is a set of tools aimed 
+at analyzing seismic records. 
+
+Concerning guided waves, many efforts are devoted to the computation of dispersion 
+curves for various waveguide geometries: Disperse [@lowe_disperse_2013] is based 
+on the global matrix method, axisym-safe-python [@kalkowski_axisymmetric_2018] 
+uses axisymmetric semi-analytical finite elements. Contact between rods and space-time
+diagrams are out of the scope of these methods.
+
+Some softwares are dedicated to the post-processing of the strain records 
+in the measuring bars of SHPB in order to obtain the stress and strain in the sample. 
+They often take dispersion into account but space-time diagrams for 1D propagation 
+are not computed [@noauthor_surepulse_2022;@francis_split_2017;@gershanik_2barg_2022].
+
 
 
 # Functionality
