+++ paper/paper.bib
@@ -1,31 +1,19 @@
 
-@article{fischer_longitudinal_1959,
-	title = {On longitudinal impact {I}. {Fundamental} cases of one-dimensional elastic impact, theories and experiments},
-	volume = {8},
-	issn = {1573-1987},
-	url = {https://doi.org/10.1007/BF00411743},
-	doi = {10.1007/BF00411743},
-	abstract = {Central longitudinal impact of slender bodies has a number of important technical applications. Theoretical treatment of these cases has, however, often been unduly scarce or crude, probably because available methods suitable for engineering applications have not been sufficiently well known. This is the first one of a series of papers intended to fill this gap by presenting theoretical solutions for a number of cases of various types, including hammers and bars of various fundamental forms, impact with elastic and plastic deformation and restraint by solid friction, and by suggesting some applications. This paper reviews known analytical, graphical and numerical methods for one-dimensional treatment of longitudinal impact and introduces a slightly modified version of the “graphodynamical” method which will be used in the following. Stress pulse measurements made with wire strain gauges on a bar impacted by cylindrical hammers of various diameters and materials are presented and found to agree reasonably well with corresponding theoretical pulse forms. Formulae and diagrams are given for the influence of the ratio of areas and material constants of hammer and bar on force, stress, energy transmission and other important quantities of the type of impact mentioned.},
-	language = {en},
-	number = {1},
-	urldate = {2021-12-08},
-	journal = {Applied Scientific Research, Section A},
-	author = {Fischer, H. C.},
-	month = jan,
-	year = {1959},
-	pages = {105--139},
-	file = {Full Text PDF:/home/dbrizard/Zotero/storage/IJGSDXGQ/Fischer - 1959 - On longitudinal impact I.pdf:application/pdf},
-}
-
-@book{johnson_impact_1972,
-	title = {Impact {Strength} of {Materials}},
-	isbn = {0-7131-3266-3},
-	publisher = {Edward Arnold},
-	author = {Johnson, W.},
-	month = jun,
-	year = {1972},
-	keywords = {IUT, {\textasciitilde}ggbook},
-	file = {JohnsonISM1972.pdf:/home/dbrizard/Zotero/storage/E9SHAKIT/JohnsonISM1972.pdf:application/pdf},
+@article{bacon_numerical_1993,
+	title = {Numerical prediction of the propagation of elastic waves in longitudinally impacted rods: {Applications} to {Hopkinson} testing},
+	volume = {13},
+	issn = {0734-743X},
+	shorttitle = {Numerical prediction of the propagation of elastic waves in longitudinally impacted rods},
+	url = {http://www.sciencedirect.com/science/article/pii/0734743X9390084K},
+	doi = {10.1016/0734-743X(93)90084-K},
[…]
