+++ paper/kalkowski.bib
@@ -1,0 +1,18 @@
+
+@article{kalkowski_axisymmetric_2018,
+	title = {Axisymmetric semi-analytical finite elements for modelling waves in buried/submerged fluid-filled waveguides},
+	volume = {196},
+	issn = {0045-7949},
+	url = {https://www.sciencedirect.com/science/article/pii/S0045794917307691},
+	doi = {10.1016/j.compstruc.2017.10.004},
+	abstract = {Efficient and accurate predictions of wave propagation are a vital component of wave-based non-destructive interrogation techniques. Although a variety of models are available in the literature, most of them are suited to a particular wave type or a specific frequency regime. In this paper we present a multi-wave model for wave propagation in axisymmetric fluid-filled waveguides, either buried or submerged in a fluid, based on the semi-analytical finite elements. The cross-section is discretised with high-order spectral elements to achieve high efficiency, and the singularities resulting from adopting a Lobatto scheme at the axis of symmetry are handled appropriately. The surrounding medium is modelled with a perfectly matched layer, and a practical rule of choice of its parameters, based only on the material properties and the geometry of the waveguide, is derived. To represent the fluid and the solid-fluid coupling, an acoustic SAFE element and appropriate coupling relationships are formulated. The model is validated against both numerical results from the literature and experiments, and the comparisons show very good agreement. Finally, an implementation of the method in Python is made available with this publication.},
+	language = {en},
+	urldate = {2022-06-07},
+	journal = {Computers \& Structures},
+	author = {Kalkowski, Michał K. and Muggleton, Jennifer M. and Rustighi, Emiliano},
+	month = feb,
+	year = {2018},
+	keywords = {Wave propagation, Semi-analytical finite element, Fluid-filled pipes, Perfectly matched layer, Spectral elements, Waves in embedded pipes, Waves in pipes},
+	pages = {327--340},
+	file = {Kalkowski et al. - 2018 - Axisymmetric semi-analytical finite elements for m.pdf:/home/dbrizard/Zotero/storage/RMSTW4W3/Kalkowski et al. - 2018 - Axisymmetric semi-analytical finite elements for m.pdf:application/pdf},
+}
