+++ paper/herrmann.bib
@@ -1,0 +1,17 @@
+
+@article{herrmann_computer_2013,
+	title = {Computer {Programs} in {Seismology}: {An} {Evolving} {Tool} for {Instruction} and {Research}},
+	volume = {84},
+	issn = {0895-0695},
+	shorttitle = {Computer {Programs} in {Seismology}},
+	url = {https://doi.org/10.1785/0220110096},
+	doi = {10.1785/0220110096},
+	number = {6},
+	urldate = {2022-06-09},
+	journal = {Seismological Research Letters},
+	author = {Herrmann, Robert B.},
+	month = nov,
+	year = {2013},
+	pages = {1081--1088},
+	file = {Herrmann - 2013 - Computer Programs in Seismology An Evolving Tool .pdf:/home/dbrizard/Zotero/storage/J3FM6TET/Herrmann - 2013 - Computer Programs in Seismology An Evolving Tool .pdf:application/pdf},
+}
