+++ paper/williams.bib
@@ -1,0 +1,21 @@
+
+@article{williams_one-dimensional_1996,
+	series = {Geophysics},
+	title = {The one-dimensional elastic wave equation: {A} finite-difference formulation for animated computer applications to full waveform propagation},
+	volume = {22},
+	issn = {0098-3004},
+	shorttitle = {The one-dimensional elastic wave equation},
+	url = {https://www.sciencedirect.com/science/article/pii/0098300495000771},
+	doi = {10.1016/0098-3004(95)00077-1},
+	abstract = {A FORTRAN computer program for modeling full waveform propagation through a layered homogeneous one-dimensional medium is presented. The output synthetic seismograms are generated using a finite-difference approximation to the wave equation, as opposed to the more conventional approach of simply convolving the calculated reflection coefficient wavetrains with representative wavelets. The strength of the program in wavefront propagation is screen animated and can be captured on hard copy. The software is user-friendly, and was developed primarily as an instructional tool. Each of the finite-difference input parameters is explained in detail for those unfamiliar with the finite-difference theory.},
+	language = {en},
+	number = {3},
+	urldate = {2022-03-18},
+	journal = {Computers \& Geosciences},
+	author = {Williams, R. Shawn and Rechtien, Richard D. and Anderson, Neil L.},
+	month = apr,
+	year = {1996},
+	keywords = {Wave propagation, Finite difference, FORTRAN, One-dimensional},
+	pages = {253--266},
+	file = {Williams et al. - 1996 - The one-dimensional elastic wave equation A finit.pdf:/home/dbrizard/Zotero/storage/2G2U2ZIN/Williams et al. - 1996 - The one-dimensional elastic wave equation A finit.pdf:application/pdf},
+}
