+++ paper/shpb.bib
@@ -1,0 +1,43 @@
+
+@article{francis_split_2017,
+	title = {Split {Hopkinson} {Pressure} {Bar} {Graphical} {Analysis} {Tool}},
+	volume = {57},
+	issn = {0014-4851, 1741-2765},
+	url = {https://link.springer.com/article/10.1007/s11340-016-0191-9},
+	doi = {10.1007/s11340-016-0191-9},
+	abstract = {An open-source Split Hopkinson Pressure Bar graphical data analysis tool has been developed. Written in Matlab®, the code can be freely distributed either as an executable binary or editable Matlab files. Beginning with raw voltages from two strain gages along with the incident/transmitted bars’ mechanical and geometrical properties, the user can visually analyze forces and displacements at the bar faces and the stresses and strains in the specimen. Wave dispersion and modulus correction are available in this package. A modest documentation and video tutorials accompany the software.},
+	language = {en},
+	number = {1},
+	urldate = {2017-03-30},
+	journal = {Experimental Mechanics},
+	author = {Francis, D. K. and Whittington, W. R. and Lawrimore, W. B. and Allison, P. G. and Turnage, S. A. and Bhattacharyya, J. J.},
+	month = jan,
+	year = {2017},
+	pages = {179--183},
+	file = {FD2SS.m:/home/dbrizard/Zotero/storage/FQMK2A34/FD2SS.m:text/x-objcsrc;francis2016.pdf:/home/dbrizard/Zotero/storage/2WA2V8SB/francis2016.pdf:application/pdf;SHPB_Analysis_Tool_Documentation.pdf:/home/dbrizard/Zotero/storage/M7NP4RJK/SHPB_Analysis_Tool_Documentation.pdf:application/pdf;SHPB_Data_Analysis.m:/home/dbrizard/Zotero/storage/E58HGRP5/SHPB_Data_Analysis.m:text/x-objcsrc},
+}
+
+@misc{gershanik_2barg_2022,
+	title = {{2BarG}},
+	url = {https://github.com/CraZMe/2BarG},
+	abstract = {2BarG – A program to process split Hopkinson (Kolsky) bar test results},
+	urldate = {2022-06-13},
+	author = {Gershanik, Tzvi},
+	month = feb,
+	year = {2022},
+	note = {original-date: 2022-02-22T08:05:40Z},
+	keywords = {kolsky, mechanics, mechanics-of-materials, mechanics-of-solids, shpb, split-hopkinson, stress-strain, stress-waves},
+	file = {Gershanik - 2022 - 2BarG.pdf:/home/dbrizard/Zotero/storage/9MAUBHXE/Gershanik - 2022 - 2BarG.pdf:application/pdf},
+}
+
+@misc{noauthor_surepulse_2022,
+	title = {{SurePulse}},
+	url = {https://github.com/relinc/SurePulse},
+	abstract = {SurePulse is a desktop application for processing sensor data from a Split Hopkinson Pressure Bar (SHPB)},
+	urldate = {2022-06-13},
+	publisher = {REL Inc.},
+	month = may,
+	year = {2022},
+	note = {original-date: 2018-07-14T15:08:18Z},
+	keywords = {java, javafx, materials-science},
+}
