+++ paper/lowe.bib
@@ -1,0 +1,11 @@
+
+@techreport{lowe_disperse_2013,
+	address = {London, UK},
+	title = {Disperse. {A} system for {Generating} {Dispersion} {Curves}. {User}'s manual},
+	number = {Version 2.0.20a},
+	institution = {Non-Destructive Testing Laboratory, Department of Mechanical Engineering, Imperial College London},
+	author = {Lowe, Michael and Pavlakovic, Brian},
+	month = jul,
+	year = {2013},
+	file = {dspv2man.pdf:/home/dbrizard/Zotero/storage/J4U3LHMD/dspv2man.pdf:application/pdf},
+}
