### 2022-06-14 #4418 @dbrizard
@MargeritPierre I updated the paper by adding a new _state of the field_ section with references to other software dealing with wave propagation.
