### 2022-06-23 ca0003c
paper updated and proofread
