+++ paper/paper.md
@@ -12,9 +12,10 @@ authors:
 affiliations:
  - name: Univ Lyon, Univ Gustave Eiffel, Univ Claude Bernard Lyon 1, LBMC UMR_T 9406, F-69622 Lyon, France
    index: 1
-date: March 2022
+date: June 2022
 bibliography: paper.bib
 ---
+<!-- LTeX: language=en-GB -->
 
 # Summary
 
@@ -45,14 +46,14 @@ is rather acceptable in numerous cases, when the length of the wave is several
 times greater than the bar radius [@johnson_impact_1972]. The material is also 
 supposed to stay in the elastic range of deformation. 
 
-# State of the field
-
-Wave propagation softwares can be found in many scientific fields. 
+To the best of our knowledge, there is no open source software available for the 
+computation of space-time diagrams, even if wave propagation software can be 
+found in many scientific fields. 
 Williams et al [@williams_one-dimensional_1996] proposed a FORTRAN program for 
 the propagation of wavefronts in layered media, this instructional tool is based 
 on finite-difference and outputs the displacement with respect to the spatial variable. 
 Computer Programs in Seismology [@herrmann_computer_2013] is a set of tools aimed 
-at analyzing seismic records. 
+at analysing seismic records. 
 
 Concerning guided waves, many efforts are devoted to the computation of dispersion 
 curves for various waveguide geometries: Disperse [@lowe_disperse_2013] is based 
@@ -60,7 +61,7 @@ on the global matrix method, axisym-safe-python [@kalkowski_axisymmetric_2018]
 uses axisymmetric semi-analytical finite elements. Contact between rods and space-time
 diagrams are out of the scope of these methods.
 
-Some softwares are dedicated to the post-processing of the strain records 
+Some pieces of software are dedicated to the post-processing of the strain records 
 in the measuring bars of SHPB in order to obtain the stress and strain in the sample. 
 They often take dispersion into account but space-time diagrams for 1D propagation 
 are not computed [@noauthor_surepulse_2022;@francis_split_2017;@gershanik_2barg_2022].
@@ -69,10 +70,10 @@ are not computed [@noauthor_surepulse_2022;@francis_split_2017;@gershanik_2barg_
 
 # Functionality
 
-The underlying equations are derived from [@bacon_numerical_1993], with a 
+The underlying equations used in `elwaspatid` are derived from [@bacon_numerical_1993], with a 
 formulation in force and velocity. The rods are discretized -between nodes- into 
 segments, each segment is supposed to have a constant impedance. The length of 
-each segments is chosen so that the transit time of the elastic wave in all the 
+each segment is chosen so that the transit time of the elastic wave in all the 
 segments is identical, which enables the use of the method of characteristics. 
 
 Force and velocity are computed at the nodes, the displacement is then deduced 
@@ -83,23 +84,23 @@ in the elements from the strain in the elements.
 ## Bar configuration, initial conditions
 
 The bar configuration can involve as many bar as needed. Each bar has constant
-properties (elastic module, cross section, density), and one-off cross-section 
+properties (elastic module, cross-section, density), and one-off cross-section 
 change can be defined. 
 
 The bars are supposed to
 be at rest (null velocity) and in contact with each other at time $t=0$. 
 
 It is possible to prescribe the initial velocity of the left bar and therefore 
-model an impactor or stricker.
-Conversly, one can provide the pulse shape arriving at the left end of the bar(s).
+model an impactor or striker.
+Conversely, one can provide the pulse shape arriving at the left end of the bar(s).
 
 ## Boundary conditions
 
-The following boundry conditions are available:
+The following boundary conditions are available:
 
 * contact interface between two bars. Compression waves can cross the interface,
   traction waves cannot;
[…]
