+++ paper/bacchini.bib
@@ -1,0 +1,20 @@
+
+@article{bocchini_graphical_2011,
+	title = {Graphical {User} {Interface} for {Guided} {Acoustic} {Waves}},
+	volume = {25},
+	issn = {1943-5487},
+	url = {https://ascelibrary.org/doi/abs/10.1061/%28ASCE%29CP.1943-5487.0000081},
+	doi = {10.1061/(ASCE)CP.1943-5487.0000081},
+	abstract = {This paper presents GUIGUW v0.1, a graphical user interface (GUI) for the computation of stress-guided wave dispersive features. The software exploits semianalytical finite-element (SAFE) formulations for the calculation of wave-propagation characteristics. The interface allows for the selection of geometrical, mechanical, and frequency-related parameters for the computation. Isotropic and anisotropic materials with linear elastic and linear viscoelastic rheological behaviors can be considered, and any waveguide cross section can be modeled. For each existing wave, the dispersive results can be represented in terms of wave number, wavelength, phase velocity, group velocity (for undamped waveguides), energy velocity, and attenuation (for damped waveguides). By simply working with the GUI, original results for guided stress waves can be obtained.},
+	language = {EN},
+	number = {3},
+	urldate = {2021-05-06},
+	journal = {Journal of Computing in Civil Engineering},
+	author = {Bocchini, Paolo and Marzani, Alessandro and Viola, Erasmo},
+	month = may,
+	year = {2011},
+	note = {Publisher: American Society of Civil Engineers},
+	keywords = {Dispersion, Acoustic techniques, Arbitrary cross section, Computation, Finite-element method, Graphical user interface, Nondestructive evaluation, Nondestructive tests, Stress waves, Stress-guided waves, Ultrasonic methods, Wave dispersion},
+	pages = {202--210},
+	file = {Bocchini et al. - 2011 - Graphical User Interface for Guided Acoustic Waves.pdf:/home/dbrizard/Zotero/storage/L9PUTWCA/Bocchini et al. - 2011 - Graphical User Interface for Guided Acoustic Waves.pdf:application/pdf},
+}
