### 2022-06-23 #4418 @diehlpk
@dbrizard I read the paper and I think it reads good. My only concern is

`We acknowledge the support from Pr Eric Jacquelin (ORCID ID) for his helpful advice on this project.`

I assume Pr is the abbreviation for professor or? It might be better to just write Dr. Eric Jacquelin or Professor Eric Jacquelin. I am not sure if Pr is a common abbreviation for most people in the world.
