### 2022-06-29 0b2e1e1
docs: fix image for readme.md
