### 2022-06-29 #4418 @diehlpk
@dbrizard Please update the title of the Zenodo archive from

`dbrizard/elwaspatid: First release after JOOS review`

to
`
elwaspatid: A Python package to compute space-time diagrams for the propagation of elastic waves in 1D rods`
