### 2022-06-29 #4418 @dbrizard
Yes, v2.0.2.

Sorry for this odd number, I would have hopped to have something like v1.0.5 but I messed things up at one point, with no possibility to go back with PyPI...
