+++ paper/paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 
 `elwaspatid` is a Python package for 1D elastic wave propagation dedicated to 
 the computation of space-time diagrams. It is meant to model the propagation of
-elastic stress waves in slender bars -or rods- generated by the impact of another bar. 
-Application can be: mechanical testing (Split Hopkinson Pressure Bars, SHPB), 
- mechanical systems which involve impact with a hammer, design of striker 
-geometry to get a specific pulse shape, understanding of wave 
-reflection/transmission in bars on academical cases.
+elastic stress waves in slender bars (or rods) generated by the impact of another bar. 
+Applications can include: mechanical testing (Split Hopkinson Pressure Bars, SHPB), 
+mechanical systems that involve impact with a hammer, design of striker 
+geometry to get a specific pulse shape, and understanding of wave 
+reflection/transmission in bars in academical cases.
 
 
 # Statement of Need
@@ -35,22 +35,22 @@ Space-time diagrams are a way to visualize the propagation of mechanical
 changes and contact interfaces can quickly become hard to compute analytically 
 [@de_saint-venant_memoire_1867;@johnson_impact_1972] or with the 
 graphodynamical method [@dejuhasz_graphical_1949;@fischer_longitudinal_1959]
- when multiple waves are combined because of wave superposition. 
+when multiple waves are combined because of wave superposition. 
 Additionally, we are often limited to simple wave shapes such as the classical
 rectangular pulse and to bars with constant cross-section; deviating from these
 conditions rapidly complicates the calculations. 
 
 The package is limited to non-dispersive propagation, which means that the pulse
 shape is not modified by the propagation, but this approximation 
-is rather acceptable in numerous cases, when the length of the wave is several
+is acceptable in numerous cases, when the length of the wave is several
 times greater than the bar radius [@johnson_impact_1972]. The material is also 
 supposed to stay in the elastic range of deformation. 
 
 To the best of our knowledge, there is no open source software available for the 
 computation of space-time diagrams, even if wave propagation software can be 
 found in many scientific fields. 
-Williams et al [@williams_one-dimensional_1996] proposed a FORTRAN program for 
-the propagation of wavefronts in layered media, this instructional tool is based 
+@williams_one-dimensional_1996 proposed a FORTRAN program for 
+the propagation of wavefronts in layered media; this instructional tool is based 
 on finite-difference and outputs the displacement with respect to the spatial variable. 
 Computer Programs in Seismology [@herrmann_computer_2013] is a set of tools aimed 
 at analysing seismic records. 
@@ -58,11 +58,11 @@ at analysing seismic records.
 Concerning guided waves, many efforts are devoted to the computation of dispersion 
 curves for various waveguide geometries: Disperse [@lowe_disperse_2013] is based 
 on the global matrix method, axisym-safe-python [@kalkowski_axisymmetric_2018] 
-uses axisymmetric semi-analytical finite elements (SAFE), 
+uses axisymmetric semi-analytical finite elements (SAFE), and 
 GUIGUW [@bocchini_graphical_2011] is also based on the SAFE formulation. 
-Contact between rods and space-time diagrams are out of the scope of these methods.
+Contact between rods and space-time diagrams are out of scope for these methods.
 
-Some pieces of software are dedicated to the post-processing of the strain records 
+Some software packages are dedicated to the post-processing of the strain records 
 in the measuring bars of SHPB in order to obtain the stress and strain in the sample. 
 They often take dispersion into account but space-time diagrams for 1D propagation 
 are not computed [@noauthor_surepulse_2022;@francis_split_2017;@gershanik_2barg_2022].
@@ -71,20 +71,20 @@ are not computed [@noauthor_surepulse_2022;@francis_split_2017;@gershanik_2barg_
 
[…]
