### 2022-07-05 #4418 @danielskatz
In addition, I've suggested a number of wording changes in https://github.com/dbrizard/elwaspatid/pull/1 - please merge this or let me know what you disagree with, then we can continue to publication
