### 2022-07-05 #4418 @dbrizard
Thank you. I am ok with your modifications., thank you for the careful proofreading.
Just note that some square brackets were dropped on some references.
Do you know if this has any consequence on these references ?
