### 2022-07-05 #4418 @danielskatz
Removing the square brackets removes the parentheses around the reference, which is preferred when the reference is being used actively rather than passively.  E.g. "Author et al. 2022 reports" vs
"in the reported literature (Author et al. 2022)"
